+++ plugins/tomcat-pool/src/main/java/is/codion/plugin/tomcat/pool/TomcatConnectionPoolFactory.java
@@ -78,6 +78,7 @@ public final class TomcatConnectionPoolFactory implements ConnectionPoolFactory 
 		@Override
 		public void close() {
 			connectionPool().close();
+			closeStatisticsCollection();
 		}
 
 		@Override